backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
pg8000==1.30.5
SQLAlchemy==2.0.23
Flask-Login==0.6.3
Werkzeug==3.0.1
